% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_raw}
\alias{df_raw}
\title{Funded projects (dataframe)}
\format{
A dataframe with 14 columns:
 \describe{
   \item{Grant}{Grant name}
   \item{Project}{Project name}
   \item{Organization}{Organization receiving grant}
   \item{Status}{Project status. Options: Ongoing, Complete}
   \item{Start_Year}{Initial funding year}
   \item{End_Year}{Final funding year}
   \item{Category}{Project category. Options: Capacity-Building, Education, 
     Implementation, Monitoring, Outreach, Planning, Research, Restoration}
   \item{Funding_Source}{Funding source. Options: BIL, NEP, SNEP}
   \item{Funding_Amount}{Amount of grant funding (US dollars)}
   \item{Report}{Link to final report}
   \item{Description}{Project description}
   \item{Latitude}{Latitude}
   \item{Longitude}{Longitude}
   \item{Popup}{Popup text for map}
 }
}
\usage{
df_raw
}
\description{
Dataframe containing project details.
}
\keyword{datasets}
