% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grantmap-package.R
\docType{package}
\name{grantmap-package}
\alias{grantmap}
\alias{grantmap-package}
\title{grantmap: NBEP funded projects}
\description{
An interactive map of projects funded by the Narragansett Bay Estuary Program.
}
\author{
\strong{Maintainer}: Mariel Sorlien \email{info@nbep.org} (\href{https://orcid.org/0000-0001-7102-2918}{ORCID})

}
\keyword{internal}
