% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_multifilter.R
\name{intersect_list}
\alias{intersect_list}
\title{Intersect list and string}
\usage{
intersect_list(x, y, delim = ",")
}
\arguments{
\item{x}{List. Input list.}

\item{y}{String. String to convert to list.}

\item{delim}{String. Delimiter used to split items in string. Default
value ",".}
}
\value{
List of matches between x and y.
}
\description{
Finds all matches between list and string. Helper function for
[multifilter()].
}
\seealso{
[multifilter()]
}
