% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_multifilter.R
\name{multifilter}
\alias{multifilter}
\title{Filter table by intersecting column values and list}
\usage{
multifilter(.data, col_name, x, delim = ",")
}
\arguments{
\item{.data}{Dataframe. Input dataframe.}

\item{col_name}{String. Name of column to check for matches.}

\item{x}{List. List to intersect with column. Rows that do not match
at least one item in list will be dropped.}

\item{delim}{String. Delimiter used to split items in string. Default
value ",".}
}
\value{
A dataframe with the same number of columns as the input. Rows that
do not meet the selection criteria have been removed, but remaining rows
remain in the same order.
}
\description{
`multifilter()` filters the input dataframe by dropping all rows where the
item(s) in the target column do not match one or more items in list.
}
\seealso{
[intersect_list()]
}
