% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_popup_text.R
\name{popup_column}
\alias{popup_column}
\title{Add column with HTML formatted popup text}
\usage{
popup_column(
  .data,
  col_name,
  col_title = NULL,
  target_col = "Popup",
  na_value = "-",
  hide_na = FALSE,
  style = "<b>in_title:</b> in_data",
  delim = "<br>"
)
}
\arguments{
\item{.data}{Dataframe.}

\item{col_name}{List or string. Columns used to populate values in popup
text. Will be added in order listed.}

\item{col_title}{List or string. The title to place at the start of each new
line. Default `NULL`.
* If list, must be same length as `col_name`.
* If `NULL`, `col_title` will be replaced with `col_name`. Underscores will
be replaced with spaces.}

\item{target_col}{String. Name of column to modify and add text to. Default
"Popup".
* If `target_col` does not exist, it will be added to the dataframe.}

\item{na_value}{String. Replacement for `NA` values. Default "-".}

\item{hide_na}{Boolean. If `TRUE`, will not add a new line of text if
`in_data` is `NA`. Default `TRUE`.}

\item{style}{String. How to format each line of data. String must include
"in_title" and "in_data" to represent the title and data value respectively.
Default "<br><b>in_title:</b> in_data".}

\item{delim}{String. Delimiter placed between each line of text. Default
"<br>".}
}
\value{
Updated dataframe with column named "Popup" unless `target_col`
set to different value. "Popup" column contains formatted popup text.
}
\description{
`popup_column` adds or updates a "Popup" column with column names and values
from the input dataframe.
* Unless `style` is updated, each column/value pair is formatted as
"<b>column name:</b> column value"
* A delimiter is added between each line of text. The default is "<br>"
}
\seealso{
[info_text()]
}
