% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_popup_text.R
\name{popup_text}
\alias{popup_text}
\title{Format paired title and value as descriptive text}
\usage{
popup_text(
  .data,
  in_title,
  in_data,
  na_value = "-",
  hide_na = FALSE,
  style = "<b>in_title:</b> in_data",
  delim = "<br>"
)
}
\arguments{
\item{.data}{String. Line of text to append new lines to.}

\item{in_title}{List or string. The title to place at the start of each new
line. If list, must be same length as `in_data`.}

\item{in_data}{List or string. Value to place after `in_title`. If list, must
be same length as `in_title`.}

\item{na_value}{String. Replacement for `NA` values. Default "-".}

\item{hide_na}{Boolean. If `TRUE`, will not add a new line of text if
`in_data` is `NA`. Default `TRUE`.}

\item{style}{String. How to format each line of data. String must include
"in_title" and "in_data" to represent the title and data value respectively.
Default "<br><b>in_title:</b> in_data".}

\item{delim}{String. Delimiter placed between each line of text. Default
"<br>".}
}
\value{
Updated string.
}
\description{
`popup_text` formats input text and values as a line of text, which it
appends to the input.
* Unless `style` is updated, each line is formatted as
"<b>`in_title`:</b> `in_data`"
* A delimiter is added between each line of text. The default is "<br>"
}
\seealso{
[popup_column()]
}
