% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shp_nbep}
\alias{shp_nbep}
\title{NBEP boundaries (shapefile)}
\format{
A dataframe with 4 columns and 3 rows:
 \describe{
   \item{Study_Area}{The name of the NBEP study area.}
   \item{DataSource}{The source(s) of the primary data used to produce this 
     dataset.}
   \item{SourceYear}{The publication year(s) of the primary data used to 
     produce this dataset.}
   \item{NBEPYear}{The publication year of the NBEP State of Narragansett 
     Bay & Its Watershed Technical Report associated with this dataset. }
   \item{geometry}{Shapefile geometry}
 }
}
\usage{
shp_nbep
}
\description{
Shapefile showing the Narragansett Bay Estuary Program's study area.
}
\keyword{datasets}
