% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dropdown.R
\name{tidy_list}
\alias{tidy_list}
\title{Tidy List}
\usage{
tidy_list(x, sort_list = TRUE, sort_decreasing = FALSE)
}
\arguments{
\item{x}{Input list.}

\item{sort_list}{Boolean. Whether to sort the list. Default TRUE.}

\item{sort_decreasing}{Boolean. Whether to sort the list in decreasing order.
Default FALSE.}
}
\value{
Updated list.
}
\description{
Sorts list and removes duplicate values. If items in list are 
  named, sorts by name.
}
