% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dropdown.R
\name{select_dropdown}
\alias{select_dropdown}
\title{Create pickerInput dropdown}
\usage{
select_dropdown(
  id,
  label,
  choices,
  choice_names = NULL,
  sort_choices = TRUE,
  decreasing = FALSE,
  add_other = TRUE,
  multiple = TRUE,
  max_options = NULL
)
}
\arguments{
\item{id}{String. Widget id.}

\item{label}{String. Widget heading/label.}

\item{choices}{List. Dropdown choices.}

\item{choice_names}{List. Display names for choice variables. If not `NULL`,
must be same length as `choices`. Default `NULL`.}

\item{sort_choices}{Boolean. If `TRUE`, sorts items in dropdown
alphabetically. If `FALSE`, does not change dropdown order and `decreasing`
argument is ignored. Default `TRUE`.}

\item{decreasing}{Boolean. If `TRUE`, sorts list in reverse alphabetical
order. If `FALSE`, sorts list in alphabetical order. Default `FALSE`.}

\item{add_other}{Boolean. If `TRUE` and `NA` value included in list, adds
"Other" to end of list. Default `TRUE`.}

\item{multiple}{Boolean. Whether to allow multiple selections. Default
`TRUE`.}

\item{max_options}{Number. Maximum number of selections. Default `NULL`.}
}
\value{
A [shinyWidgets::pickerInput()] widget.
}
\description{
`select_dropdown()` creates a [shinyWidgets::pickerInput()]
dropdown widget.
}
